requirements.txt
transformers
sentence-transformers
numpy
faiss-cpu
pandas
tqdm
PyPDF2
python-docx